Enemy's shoots
Enemy now can shoot, but it does not kill the player. That is because the enemy function is executed before the player function. In order to correct that I need to separate both functions into little functions and execute the shoots of the enemy before executing the function which processes the death of the player ship.