Game optimized part 2
Terrain drawing was moved to the main loop to optimize more the SysTick interrupt.
Also times were measured to determine if the system is real time. The SysTick interrupt occurs exactly every 33.25 ms (30 Hz), The SysTick interrupt takes 232 us and the main loop takes 2.327 ms. That makes a total execution time of 2.568 ms and a "wait free time" of 30.682 ms. So the game could be way more complicated and the processor could execute it. Or the frequency of the processor could be eventually be decreased to save energy.