Enemy updated
Now before a new enemy is created, it is checked if the pixels where the enemy will be created are already occupied. If that is the case, the enemy is not created in that cycle.

Also the enemy can move now if it is in a corner. Before, it would only move if it was in the center of the display.